+++ core/src/test/kotlin/net/corda/core/contracts/AmountTests.kt
@@ -1,6 +1,7 @@
 package net.corda.core.contracts
 
 import net.corda.finance.*
+import net.corda.core.contracts.Amount.Companion.sumOrZero
 import org.junit.Test
 import java.math.BigDecimal
 import java.util.*
