+++ docs/packages.md
@@ -1,0 +1,3 @@
+# Package net.corda.finance.utils
+
+A collection of utilities for summing financial states, for example, summing obligations to get total debts.
