+++ docs/source/changelog.rst
@@ -45,6 +45,12 @@ UNRELEASED
 * Remove `IssuerFlow` as it allowed nodes to request arbitrary amounts of cash to be issued from any remote node. Use
   `CashIssueFlow` instead.
 
+* Some utility/extension functions (``sumOrThrow``, ``sumOrNull``, ``sumOrZero`` on ``Amount`` and ``Commodity``)
+  have moved to be static methods on the classes themselves. This improves the API for Java users who no longer
+  have to see or known about file-level FooKt style classes generated by the Kotlin compile, but means that IntelliJ
+  no longer auto-suggests these extension functions in completion unless you add import lines for them yourself
+  (this is Kotlin IDE bug KT-15286).
+
 Milestone 14
 ------------
